Refactor `proto::message::Identity` into module
